Exception

+++ Kouvee/View/Data/Hapus/FormHapusHewan.cs
@@ -92,6 +92,12 @@ namespace Kouvee.View.Data.Hapus
         {
             try
             {
+                if (string.IsNullOrEmpty(txtCari.Text.Trim()))
+                {
+                    MessageBox.Show("Text Pencarian Kosong");
+                    throw null;
+                }
+
                 var list = new HewanControl();
                 list.DeleteHewan(txtCari.Text);
             }
