+++ Kouvee/View/Data/Hapus/FormHapusPegawai.cs
@@ -51,6 +51,12 @@ namespace Kouvee.View.Data.Hapus
         {
             try
             {
+                if (string.IsNullOrEmpty(txtCari.Text.Trim()))
+                {
+                    MessageBox.Show("Text Pencarian Kosong");
+                    throw null;
+                }
+
                 var list = new PegawaiControl();
                 list.DeletePegawai(txtCari.Text);
             }
